Refactor/37 db 데이터 정제 및 enum 저장 방식 개선

+++ db/models/criterion/general_criterion.py
@@ -1,5 +1,5 @@
 # db/models/criterion/general_criterion.py
-from sqlalchemy import Column, Integer, ForeignKey, TEXT, TIMESTAMP, text
+from sqlalchemy import Column, BIGINT, Integer, ForeignKey, TEXT, TIMESTAMP, text
 from db.database import Base
 import json
 
@@ -7,8 +7,8 @@ class GeneralCriterion(Base):
     __tablename__ = "general_criterion"
     __table_args__ = {'extend_existing': True}
     
-    id = Column(Integer, primary_key=True, index=True)
-    scholarship_id = Column(Integer, ForeignKey("scholarship.id"), comment="Scholarship ID (PK, FK)")
+    id = Column(BIGINT, primary_key=True, index=True)
+    scholarship_id = Column(BIGINT, ForeignKey("scholarship.id"), comment="Scholarship ID (PK, FK)")
     
     required_qualifications = Column(TEXT, comment="장학금 지원을 위한 필수 자격 조건 목록 (JSON 문자열로 저장)")
     preference_qualifications = Column(TEXT, comment="우대 조건 목록 (JSON 문자열로 저장)")
